fixed wrong return type of getMaxFFTSize

+++ src/de/sciss/dsp/ConstQ.java
@@ -115,7 +115,7 @@ public class ConstQ
 		this.maxFFTSize	= maxFFTSize;
 	}
 	
-	public float getMaxFFTSize()
+	public int getMaxFFTSize()
 	{
 		return maxFFTSize;
 	}
